+++ public/change_editor_option_event.js
@@ -1,0 +1,6 @@
+export class ChangeEditorOptionEvent extends CustomEvent {
+    
+    constructor(name, value) {
+        super('Editor:changeoption', {detail: {name: name, value: value}});
+    }
+}